refactor: extract context filesystem fact updates
